dashboard: refactor authorize page
